requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.94.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-secret-manager==2.16.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.48.2
h3==3.7.6
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
langchain==0.0.235
langsmith==0.0.7
loguru==0.7.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.5.1
predicthq
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.1
PyYAML==6.0
pyzmq==25.0.2
replicate==0.8.3
requests==2.30.0
rich==13.3.5
rsa==4.9
schematics==2.0.0.dev2
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.18
stack-data==0.6.2
streamlit==1.22.0
tenacity==8.2.2
timezonefinder==6.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
langchain-openai
pydantic
